+++ data/graphic/shaders/fs_lightprobe_diffuse_sampling.glsl
@@ -51,11 +51,11 @@ vec3 GetImportanceSampleDiffuseFrostbite(in vec3 _NormalReflection)
     float NdotL;
     float PDF;
     
-    const uint NumSamples = NUM_SAMPLES;
+    const int NumSamples = NUM_SAMPLES;
     
     Result = vec3(0.0f);
     
-    for(uint IndexOfSample = 0; IndexOfSample < NumSamples; ++IndexOfSample )
+    for(int IndexOfSample = 0; IndexOfSample < NumSamples; ++IndexOfSample )
     {
         Xi = GetHammersley(IndexOfSample, NumSamples);
         
@@ -65,11 +65,11 @@ vec3 GetImportanceSampleDiffuseFrostbite(in vec3 _NormalReflection)
         
         if (NdotL > 0.0f)
         {
-            Result += textureLod(ps_EnvironmentCubemap, L, 0).rgb;
+            Result += textureLod(ps_EnvironmentCubemap, L, 0.0f).rgb;
         }
     }
     
-    return Result * (1.0f / NumSamples);
+    return Result * (1.0f / float(NumSamples));
 }
 
 // -----------------------------------------------------------------------------
@@ -79,11 +79,11 @@ vec3 GetImportanceSampleDiffuseMipmapped(in vec3 _NormalReflection)
     vec3 V = _NormalReflection;
     vec4 Result = vec4(0.0f);
     
-    const uint NumSamples = NUM_SAMPLES;
+    const int NumSamples = NUM_SAMPLES;
     
     ivec2 CubeSize = textureSize(ps_EnvironmentCubemap, 0);
     
-    for(uint IndexOfSample = 0; IndexOfSample < NumSamples; ++IndexOfSample )
+    for(int IndexOfSample = 0; IndexOfSample < NumSamples; ++IndexOfSample )
     {
         vec2  Xi  = GetHammersley(IndexOfSample, NumSamples);
         vec3  H   = GetImportanceSampleCosDirection(Xi, _NormalReflection);
@@ -97,8 +97,8 @@ vec3 GetImportanceSampleDiffuseMipmapped(in vec3 _NormalReflection)
         // -----------------------------------------------------------------------------
         float PDF = max(0.0, dot(_NormalReflection, L) * INV_PI);
         
-        float SolidAngleTexel  = 4.0f * PI / (6.0f * CubeSize.x * CubeSize.y);
-        float SolidAngleSample = 1.0f / (NumSamples * PDF);
+        float SolidAngleTexel  = 4.0f * PI / (6.0f * float(CubeSize.x * CubeSize.y));
+        float SolidAngleSample = 1.0f / (float(NumSamples) * PDF);
         
         //TODO: add mip count to constantbuffer
         float LOD = 0.5f * log2(SolidAngleSample / SolidAngleTexel);
